% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{GeometricMorphometricsMix-defunct}
\alias{GeometricMorphometricsMix-defunct}
\title{Defunct functions in GeometricMorphometricsMix}
\description{
These functions are defunct and have been removed from the package.
}
\section{Defunct functions}{

\describe{
  \item{\code{Kmultparallel()}}{Defunct.}
  \item{\code{print.parallel_Kmult()}}{Defunct along with
    \code{Kmultparallel()}.}
  \item{\code{plot.parallel_Kmult()}}{Defunct along with
    \code{Kmultparallel()}.}
  \item{\code{summary.parallel_Kmult()}}{Defunct along with
    \code{Kmultparallel()}.}
}
}

\keyword{internal}
