% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EscoufierRVrarefy.R
\name{plot.EscoufierRVrarefy}
\alias{plot.EscoufierRVrarefy}
\title{Plot method for EscoufierRVrarefy objects}
\usage{
\method{plot}{EscoufierRVrarefy}(x, point_color = "darkblue", errorbar_color = "darkred", ...)
}
\arguments{
\item{x}{An object of class "EscoufierRVrarefy"}

\item{point_color}{A single color or a vector of colors for point estimates.
If length 1, the same color is used for all points. If length equals the
number of groups, colors are assigned per group. (default "darkblue")}

\item{errorbar_color}{A single color or a vector of colors for error bars.
Follows the same recycling rules as `point_color`. (default "darkred")}

\item{...}{Additional arguments passed to the underlying plotting function}
}
\value{
A ggplot object
}
\description{
Creates a confidence interval plot for RV rarefaction results
}
