% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmultparallel.R
\name{plot.parallel_Kmult}
\alias{plot.parallel_Kmult}
\title{Plot method for parallel_Kmult objects (defunct)}
\usage{
\method{plot}{parallel_Kmult}(x, alpha = 0.25, title = NULL, x_lab = "Kmult", ...)
}
\arguments{
\item{x}{Ignored (function is defunct).}

\item{alpha}{Ignored (function is defunct).}

\item{title}{Ignored (function is defunct).}

\item{x_lab}{Ignored (function is defunct).}

\item{...}{Ignored (function is defunct).}
}
\value{
Does not return; always signals a defunct error.
}
\description{
**Defunct**: Removed from the package along with
\code{Kmultparallel()}.
}
\keyword{internal}
