% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_wregr.R
\name{plot.phylo_wregression}
\alias{plot.phylo_wregression}
\title{Plot method for phylo_wregression objects}
\usage{
\method{plot}{phylo_wregression}(x, plot_phylogeny = TRUE, ...)
}
\arguments{
\item{x}{An object of class `phylo_wregression`, 
typically the result of \code{phylo_wregression()}.}

\item{plot_phylogeny}{Logical; whether to include the phylogeny in plots.
If TRUE and simulation results are present in the `x` object, tips exceeding the range of simulated values are highlighted on the phylogeny.
If FALSE the plot will be for multivariate data a scatterplot along the first two principal components of coefficients
(highlighting the ones which exceed the simulated threshold if simulations were carried out),
for univariate data a density plot.}

\item{...}{Additional arguments passed to the underlying plotting functions (e.g., graphical parameters).}
}
\value{
Invisibly returns the input object `x`.
}
\description{
Visualizes the results of phylogenetically weighted regression analysis,
highlighting tips with coefficients or angles that exceed simulation-based thresholds.
}
\note{
The `phytools` package is required for phylogenetic plotting and stochastic mapping.
If not installed, only non-phylogenetic plots are available.
}
\examples{
# Simulate a tree and traits
library(phytools)
library(mvMORPH)
set.seed(123)
tree = pbtree(n=30)
Sigma=diag(6); Sigma[1,2]=Sigma[2,1]=0.5
Data_sim=mvSIM(tree=tree, nsim=1, model="BM1", param=list(sigma=Sigma))
Y = Data_sim[,2:6]
X = Data_sim[,1]
result = phylo_wregression(tree, Y, X, nsim=100)
plot(result)

}
