% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity_resample.R
\name{print.disparity_resample}
\alias{print.disparity_resample}
\title{Print method for disparity_resample objects}
\usage{
\method{print}{disparity_resample}(x, ...)
}
\arguments{
\item{x}{An object of class "disparity_resample"}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Prints results table and checks for CI overlap among groups
}
