% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EscoufierRVrarefy.R
\name{print.EscoufierRVrarefy}
\alias{print.EscoufierRVrarefy}
\title{Print method for EscoufierRVrarefy objects}
\usage{
\method{print}{EscoufierRVrarefy}(x, ...)
}
\arguments{
\item{x}{An object of class "EscoufierRVrarefy"}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Prints results table and checks for CI overlap among groups
}
