% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RarefiedDisparity.R
\name{RarefiedDisparity}
\alias{RarefiedDisparity}
\title{Rarefied estimates of disparity}
\usage{
RarefiedDisparity(Data, rep = 1000, samplesize)
}
\arguments{
\item{Data}{Matrix or data frame containing data
(observations in rows, variables in columns).}

\item{rep}{number of resamplings to obtain the rarefied estimate}

\item{samplesize}{sample size to which the rarefaction procedure is carried out}
}
\value{
The function outputs a list with the following elements:
 \describe{
  \item{RarefiedSamplesEstimates}{Estimates of both multivariate variance and
        mean pairwise Euclidean distance for each of the rarefied samples}
  \item{Descriptives}{Mean, standard deviation, median and 2.5\% (Min) and 97.5\% (Max) across the rarefied samples}
}
}
\description{
Computes rarefied estimates of multivariate variance
(i.e., the trace of the covariance matrix) and
the mean pairwise Euclidean distance
}
\details{
The advantage of computing rarefied estimates of a parameter
is to account for differences in sample sizes.
This can be useful to compare estimates among groups.

This function is similar to the approaches implmented in the
Matlab package MDA (Navarro 2003).

Examples of these approaches abound in the literature
(e.g., Drake & Klingenberg 2010; Fruciano et al. 2014, 2016)
}
\section{Notice}{

the function does NOT perform GPA on each rarefied sample
this may or may not make a difference in estimates.
In many cases, it will probably not make much difference
}

\references{
Navarro N. 2003. MDA: a MATLAB-based program for morphospace-disparity analysis. Computers & Geosciences 29:655-664.

Drake AG, Klingenberg CP. 2010. Large-scale diversification of skull shape in domestic dogs: disparity and modularity. American Naturalist 175:289-301.

Fruciano C, Franchini P, Raffini F, Fan S, Meyer A. 2016. Are sympatrically speciating Midas cichlid fish special? Patterns of morphological and genetic variation in the closely related species Archocentrus centrarchus. Ecology and Evolution 6:4102-4114.

Fruciano C, Pappalardo AM, Tigano C, Ferrito V. 2014. Phylogeographical relationships of Sicilian brown trout and the effects of genetic introgression on morphospace occupation. Biological Journal of the Linnean Society 112:387-398.
}
