% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{GeometricMorphometricsMix-defunct}
\alias{GeometricMorphometricsMix-defunct}
\title{Defunct functions in GeometricMorphometricsMix}
\description{
These functions are defunct and have been removed from the package.
Please use the modern alternatives as indicated below.
}
\keyword{internal}
