% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{GeometricMorphometricsMix-defunct}
\alias{GeometricMorphometricsMix-defunct}
\alias{rarefied_convex_hull}
\alias{rarefied_disparity}
\title{Defunct functions in GeometricMorphometricsMix}
\usage{
rarefied_convex_hull(Data, rep = 100, samplesize)

rarefied_disparity(Data, rep = 1000, samplesize, claramunt_pv = FALSE)
}
\description{
These functions are defunct and have been removed from the package.
Please use the modern alternatives as indicated below.
}
\section{\code{rarefied_convex_hull}}{

For \code{rarefied_convex_hull}, use \code{\link{disparity_resample}} with 
\code{statistic='convex_hull_volume'} instead.
}

\section{\code{rarefied_disparity}}{

For \code{rarefied_disparity}, use \code{\link{disparity_resample}} with 
appropriate statistic parameter instead.
}

\keyword{internal}
