% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{GeometricMorphometricsMix-deprecated}
\alias{GeometricMorphometricsMix-deprecated}
\title{Deprecated functions in GeometricMorphometricsMix}
\description{
These functions are deprecated and will be removed in a future version
of the package. Please migrate to the suggested alternatives.
}
\section{Deprecated functions}{

\describe{
  \item{\code{Kmultparallel()}}{Deprecated.}
  \item{\code{print.parallel_Kmult()}}{Deprecated along with
    \code{Kmultparallel()}.}
  \item{\code{plot.parallel_Kmult()}}{Deprecated along with
    \code{Kmultparallel()}.}
  \item{\code{summary.parallel_Kmult()}}{Deprecated along with
    \code{Kmultparallel()}.}
}
}

\keyword{internal}
