% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparity_resample.R
\name{plot.disparity_resample}
\alias{plot.disparity_resample}
\title{Plot method for disparity_resample objects}
\usage{
\method{plot}{disparity_resample}(x, ...)
}
\arguments{
\item{x}{An object of class "disparity_resample"}

\item{...}{Additional arguments}
}
\value{
A ggplot object
}
\description{
Creates a confidence interval plot for disparity resample results
}
